chore: rename lib/io/filesystem => lib/io/Filesystem
