+++ src/lib/io/Filesystem.cpp
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: GPL-2.0-only WITH LicenseRef-OpenSSL-Exception
  */
 
-#include "filesystem.h"
+#include "Filesystem.h"
 
 #include "common/common.h"
