+++ src/lib/net/FingerprintDatabase.h
@@ -9,7 +9,7 @@
 
 #include "FingerprintData.h"
 
-#include "io/filesystem.h"
+#include "io/Filesystem.h"
 
 #include <iosfwd>
 #include <string>
