+++ src/lib/net/SecureSocket.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "io/filesystem.h"
+#include "io/Filesystem.h"
 #include "net/SecurityLevel.h"
 #include "net/TCPSocket.h"
 #include "net/XSocket.h"
