+++ src/lib/net/SecureUtils.cpp
@@ -9,7 +9,7 @@
 
 #include "base/FinalAction.h"
 #include "base/String.h"
-#include "io/filesystem.h"
+#include "io/Filesystem.h"
 
 #include <openssl/evp.h>
 #include <openssl/pem.h>
